+++ tests/app/NestedObjects/City.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Spiral\Serializer\Symfony\Tests\App\NestedObjects;
+
+final class City implements \JsonSerializable
+{
+    public function __construct(
+        public string $name,
+        public \DateTimeZone $timezone,
+    ) {
+    }
+
+    public function jsonSerialize(): array
+    {
+        return [
+            'name' => $this->name,
+            'timezone' => $this->timezone->getName(),
+        ];
+    }
+}