+++ tests/src/Feature/SerializerTest.php
@@ -6,6 +6,8 @@ namespace Spiral\Serializer\Symfony\Tests\Feature;
 
 use Spiral\Serializer\SerializerManager;
 use Spiral\Serializer\Symfony\Serializer;
+use Spiral\Serializer\Symfony\Tests\App\NestedObjects\City;
+use Spiral\Serializer\Symfony\Tests\App\NestedObjects\Country;
 use Spiral\Serializer\Symfony\Tests\App\Object\Post;
 use Spiral\Serializer\Symfony\Tests\App\Object\Product;
 
@@ -23,7 +25,7 @@ final class SerializerTest extends TestCase
     {
         $manager = $this->getContainer()->get(SerializerManager::class);
 
-        $result = $manager->unserialize('{"id":1,"text":"some","active":false,"views":3}', Post::class, 'json');
+        $result = $manager->unserialize('{"id":1,"text":"some","active":false,"views":3}', Post::class, 'symfony-json');
         $this->assertInstanceOf(Post::class, $result);
         $this->assertSame(1, $result->id);
         $this->assertSame('some', $result->text);
@@ -32,7 +34,7 @@ final class SerializerTest extends TestCase
         $result = $manager->unserialize(
             '{"id":1,"text":"some","active":false,"views":3}',
             new Post(2, '', true, 1),
-            'json'
+            'symfony-json'
         );
         $this->assertInstanceOf(Post::class, $result);
         $this->assertSame(1, $result->id);
@@ -47,7 +49,7 @@ final class SerializerTest extends TestCase
         $result = $manager->unserialize(
             '{"id":1,"title":"Some product","price":100,"active":false,"product_views":5}',
             Product::class,
-            'json'
+            'symfony-json'
         );
         $this->assertInstanceOf(Product::class, $result);
         $this->assertSame(1, $result->id);
@@ -57,11 +59,27 @@ final class SerializerTest extends TestCase
         $this->assertSame(5, $result->views);
     }
 
+    public function testUnserializeNestedObjects(): void
+    {
+        $manager = $this->getContainer()->get(SerializerManager::class);
+
+        $result = $manager->unserialize(
+            '{"name":"USA","cities":[{"name":"Chicago","timezone":"America\/Chicago"},{"name":"NewYork","timezone":"America\/New_York"}]}',
+            Country::class,
+            'symfony-json'
+        );
+
+        $this->assertInstanceOf(Country::class, $result);
+        $this->assertSame('USA', $result->name);
+        $this->assertSame('Chicago', $result->cities[0]->name);
+        $this->assertSame('NewYork', $result->cities[1]->name);
+    }
+
     public function testGroupNormalize(): void
     {
         $manager = $this->getContainer()->get(SerializerManager::class);
         /** @var Serializer $serializer */
-        $serializer = $manager->getSerializer('json');
+        $serializer = $manager->getSerializer('symfony-json');
 
         $product = new Product(1, 'Some product', 100, false, 5);
 
@@ -76,20 +94,44 @@ final class SerializerTest extends TestCase
         $this->assertSame(['product_views' => 5], $serializer->normalize($product, null, ['groups' => 'group3']));
     }
 
-    public function serializeDataProvider(): \Traversable
+    public static function serializeDataProvider(): \Traversable
     {
-        yield ['{"id":1,"text":"some","active":false,"views":3}', new Post(1, 'some', false, 3), 'json'];
-        yield ['id,text,active,views1,some,1,5', new Post(1, 'some', true, 5), 'csv'];
-        yield ['{id:1,text:some,active:true,views:5}', new Post(1, 'some', true, 5), 'yaml'];
+        yield ['{"id":1,"text":"some","active":false,"views":3}', new Post(1, 'some', false, 3), 'symfony-json'];
+        yield ['id,text,active,views1,some,1,5', new Post(1, 'some', true, 5), 'symfony-csv'];
+        yield ['{id:1,text:some,active:true,views:5}', new Post(1, 'some', true, 5), 'symfony-yaml'];
         yield [
             '<?xmlversion="1.0"?><response><id>1</id><text>some</text><active>1</active><views>5</views></response>',
             new Post(1, 'some', true, 5),
-            'xml'
+            'symfony-xml'
         ];
         yield [
             '{"id":1,"title":"Someproduct","price":100.0,"active":false,"product_views":5}',
             new Product(1, 'Some product', 100, false, 5),
-            'json'
+            'symfony-json'
+        ];
+        yield [
+            '{"name":"USA","cities":[{"name":"Chicago","timezone":"America\/Chicago"},{"name":"NewYork","timezone":"America\/New_York"}]}',
+            new Country('USA', [
+                new City('Chicago', new \DateTimeZone('America/Chicago')),
+                new City('New York', new \DateTimeZone('America/New_York'))
+            ]),
+            'symfony-json'
+        ];
+        yield [
+            'name,cities.0.name,cities.0.timezone,cities.1.name,cities.1.timezoneUSA,Chicago,America/Chicago,"NewYork",America/New_York',
+            new Country('USA', [
+                new City('Chicago', new \DateTimeZone('America/Chicago')),
+                new City('New York', new \DateTimeZone('America/New_York'))
+            ]),
+            'symfony-csv'
+        ];
+        yield [
+            '<?xmlversion="1.0"?><response><name>USA</name><cities><name>Chicago</name><timezone>America/Chicago</timezone></cities><cities><name>NewYork</name><timezone>America/New_York</timezone></cities></response>',
+            new Country('USA', [
+                new City('Chicago', new \DateTimeZone('America/Chicago')),
+                new City('New York', new \DateTimeZone('America/New_York'))
+            ]),
+            'symfony-xml'
         ];
     }
 }